Add OllamaTeacher, Qwen3 integration tests, and fix GPU OOM issues
   - Add genesis/models/ollama_teacher.py: Ollama-backed teacher for KD
     using network Qwen3.5:cloud, no local GPU memory required
   - Add tests/test_qwen3_integration.py: 21 real integration tests covering
     OllamaTeacher, Qwen3-1.7B student (LoRA), KD loss, distillation trainer,
     and population evolution, all passing against network Ollama server
   - Fix genetics.py crossover/mutate to run on CPU to prevent GPU OOM
   - Fix teacher.py to use device_map=auto for multi-GPU model loading
   - Fix torch.load calls to pass weights_only=False throughout
   - Add Qwen3 model variants to scripts/download_models.py
   - Fix HF token detection to check both old and new cache locations
   - Expand test_fitness.py with QAFitness unit tests
   - Add test_datasets.py, test_integration.py, test_optimizer.py, test_tts.py

+++ genesis/models/student.py
@@ -73,7 +73,7 @@ class StudentModel:
         # Load base model
         self._model = AutoModelForCausalLM.from_pretrained(
             self.model_name_or_path,
-            torch_dtype=self.dtype,
+            dtype=self.dtype,
             device_map=self.device if quantization_config else None,
             quantization_config=quantization_config,
             trust_remote_code=self.trust_remote_code,
